Remove development dependency of guard-rspec
Down the dependecy tree we go!

This requires `listen` which requires `ruby_dep`. `ruby_dep` requires a
minimum Ruby version of 2.2.3 which means that we are unable to run our
tests under Ruby 2.1 which is causing our builds to fail.

This gem does not provide enough value to stay as a dependency and
removing it doesn't stop a developer from using it locally if they want.