Use to_json instead of template
Generally, we would prefer using a template to calling to_json on an
object. However this template simply dumped ALL columns of the relation
record, so it would be simpler to just use to_json.

This also avoids rewriting the template in Jbuilder, which would be fine
and easy to do and not a problem, but since this isn't really a template
to begin with (dumping all attributes), we might as well not.