Rename cache_key interface methods
In this commit:
https://github.com/solidusio-contrib/solidus_related_products/commit/127ce66a10ddbf8d55e339a0823261f53b885172
we implement cache_key interface methods to invalidate the variant cache
when a product is modified since the related variant is available when
the product is available.

We didn't notice that the cache_key method already exist into the Rails
APIs. Overriding it could create some problems in production.