Add Bi-Directional boolean to RelationType model
This commit adds the bidirectional column to the table spree_relation_types
and changes the backend to allow the admin users to manage it.
The Bi-Directional can only be set if the RelationType is applied to the
same models and can be set only on creation, this is needed to avoid
unpredictable behavior.