Add support for Zeitwerk
This commit adds the support for the Zeitwerk code loader by using
SolidusSupport::EngineExtensions instead of SolidusSupport::EngineExtensions::Decorators
and removing the custom loader for the decorators in the engine.
It also moves the decorator files to be correctly added to the autoloader
and also be loaded on engine boot.

For more information, see:

https://github.com/solidusio/solidus_support#engine-extensions