Add :no_conflict configuration option
If the :no_conflict configuration is set to true, then the
method_missing in the product and variant decorators is disabled.
This is to appease developers that think it is dangerous to
allow end-users to define metaprogrammatic behavior in their
application.

By default, :no_conflict is set to false so as not to break any
existing apps.