Ensure `@relation_types` is assigned before #update
When the admin variants controller cannot successfully complete an
`#update` action, it re-renders the user back on the `#edit` action
view.

This causes `@relation_types` to be nil and makes the Deface template
that tries to load product relations error out on a
`@relation_types.empty?` block.

We can avoid this by making sure this before action is performed on
`#update`, too.

Co-Authored-By: Adam Mueller <adam@super.gd>